refactor device managers
* moves the implementation of the DeviceManagers away from the devices themselves (dev -> pltf)
* makes acccess to the platforms uniform to all other types in alpaka by introducing traits (DevManCpu::getDevByIdx(0) -> alpaka::pltf::getDevByIdx<DevManCpu>(0u))
