+++ include/alpaka/acc/AccCpuFibers.hpp
@@ -38,6 +38,7 @@
 #include <alpaka/acc/Traits.hpp>                // acc::traits::AccType
 #include <alpaka/dev/Traits.hpp>                // dev::traits::DevType
 #include <alpaka/exec/Traits.hpp>               // exec::traits::ExecType
+#include <alpaka/pltf/Traits.hpp>               // pltf::traits::PltfType
 #include <alpaka/size/Traits.hpp>               // size::traits::SizeType
 
 // Implementation details.
@@ -252,17 +253,6 @@ namespace alpaka
             {
                 using type = dev::DevCpu;
             };
-            //#############################################################################
-            //! The CPU fibers accelerator device type trait specialization.
-            //#############################################################################
-            template<
-                typename TDim,
-                typename TSize>
-            struct DevManType<
-                acc::AccCpuFibers<TDim, TSize>>
-            {
-                using type = dev::DevManCpu;
-            };
         }
     }
     namespace dim
@@ -300,6 +290,23 @@ namespace alpaka
                 TArgs...>
             {
                 using type = exec::ExecCpuFibers<TDim, TSize, TKernelFnObj, TArgs...>;
+            };
+        }
+    }
+    namespace pltf
+    {
+        namespace traits
+        {
+            //#############################################################################
+            //! The CPU fibers executor platform type trait specialization.
+            //#############################################################################
+            template<
+                typename TDim,
+                typename TSize>
+            struct PltfType<
+                acc::AccCpuFibers<TDim, TSize>>
+            {
+                using type = pltf::PltfCpu;
             };
         }
     }
