+++ include/alpaka/acc/AccCpuOmp4.hpp
@@ -42,6 +42,7 @@
 #include <alpaka/acc/Traits.hpp>                // acc::traits::AccType
 #include <alpaka/exec/Traits.hpp>               // exec::traits::ExecType
 #include <alpaka/dev/Traits.hpp>                // dev::traits::DevType
+#include <alpaka/pltf/Traits.hpp>                   // pltf::traits::PltfType
 #include <alpaka/size/Traits.hpp>               // size::traits::SizeType
 
 // Implementation details.
@@ -239,17 +240,6 @@ namespace alpaka
             {
                 using type = dev::DevCpu;
             };
-            //#############################################################################
-            //! The CPU OpenMP 4.0 accelerator device type trait specialization.
-            //#############################################################################
-            template<
-                typename TDim,
-                typename TSize>
-            struct DevManType<
-                acc::AccCpuOmp4<TDim, TSize>>
-            {
-                using type = dev::DevManCpu;
-            };
         }
     }
     namespace dim
@@ -287,6 +277,23 @@ namespace alpaka
                 TArgs...>
             {
                 using type = exec::ExecCpuOmp4<TDim, TSize, TKernelFnObj, TArgs...>;
+            };
+        }
+    }
+    namespace pltf
+    {
+        namespace traits
+        {
+            //#############################################################################
+            //! The CPU OpenMP 4.0 executor platform type trait specialization.
+            //#############################################################################
+            template<
+                typename TDim,
+                typename TSize>
+            struct PltfType<
+                acc::AccCpuOmp4<TDim, TSize>>
+            {
+                using type = pltf::PltfCpu;
             };
         }
     }
