+++ include/alpaka/acc/AccCpuThreads.hpp
@@ -38,6 +38,7 @@
 #include <alpaka/acc/Traits.hpp>                    // acc::traits::AccType
 #include <alpaka/dev/Traits.hpp>                    // dev::traits::DevType
 #include <alpaka/exec/Traits.hpp>                   // exec::traits::ExecType
+#include <alpaka/pltf/Traits.hpp>                   // pltf::traits::PltfType
 #include <alpaka/size/Traits.hpp>                   // size::traits::SizeType
 
 // Implementation details.
@@ -251,17 +252,6 @@ namespace alpaka
             {
                 using type = dev::DevCpu;
             };
-            //#############################################################################
-            //! The CPU threads accelerator device type trait specialization.
-            //#############################################################################
-            template<
-                typename TDim,
-                typename TSize>
-            struct DevManType<
-                acc::AccCpuThreads<TDim, TSize>>
-            {
-                using type = dev::DevManCpu;
-            };
         }
     }
     namespace dim
@@ -299,6 +289,23 @@ namespace alpaka
                 TArgs...>
             {
                 using type = exec::ExecCpuThreads<TDim, TSize, TKernelFnObj, TArgs...>;
+            };
+        }
+    }
+    namespace pltf
+    {
+        namespace traits
+        {
+            //#############################################################################
+            //! The CPU threads executor platform type trait specialization.
+            //#############################################################################
+            template<
+                typename TDim,
+                typename TSize>
+            struct PltfType<
+                acc::AccCpuThreads<TDim, TSize>>
+            {
+                using type = pltf::PltfCpu;
             };
         }
     }
