+++ include/alpaka/alpaka.hpp
@@ -197,6 +197,13 @@
 #include <alpaka/offset/Traits.hpp>
 
 //-----------------------------------------------------------------------------
+// platform
+//-----------------------------------------------------------------------------
+#include <alpaka/pltf/PltfCpu.hpp>
+#include <alpaka/pltf/PltfCudaRt.hpp>
+#include <alpaka/pltf/Traits.hpp>
+
+//-----------------------------------------------------------------------------
 // rand
 //-----------------------------------------------------------------------------
 #include <alpaka/rand/RandCuRand.hpp>
