+++ include/alpaka/exec/ExecGpuCudaRt.hpp
@@ -32,6 +32,7 @@
 #include <alpaka/dev/Traits.hpp>                // dev::traits::DevType
 #include <alpaka/dim/Traits.hpp>                // dim::traits::DimType
 #include <alpaka/exec/Traits.hpp>               // exec::traits::ExecType
+#include <alpaka/pltf/Traits.hpp>               // pltf::traits::PltfType
 #include <alpaka/size/Traits.hpp>               // size::traits::SizeType
 #include <alpaka/stream/Traits.hpp>             // stream::traits::Enqueue
 
@@ -200,19 +201,6 @@ namespace alpaka
             {
                 using type = dev::DevCudaRt;
             };
-            //#############################################################################
-            //! The GPU CUDA executor device manager type trait specialization.
-            //#############################################################################
-            template<
-                typename TDim,
-                typename TSize,
-                typename TKernelFnObj,
-                typename... TArgs>
-            struct DevManType<
-                exec::ExecGpuCudaRt<TDim, TSize, TKernelFnObj, TArgs...>>
-            {
-                using type = dev::DevManCudaRt;
-            };
         }
     }
     namespace dim
@@ -252,6 +240,25 @@ namespace alpaka
                 TArgs...>
             {
                 using type = exec::ExecGpuCudaRt<TDim, TSize, TKernelFnObj, TArgs...>;
+            };
+        }
+    }
+    namespace pltf
+    {
+        namespace traits
+        {
+            //#############################################################################
+            //! The CPU CUDA executor platform type trait specialization.
+            //#############################################################################
+            template<
+                typename TDim,
+                typename TSize,
+                typename TKernelFnObj,
+                typename... TArgs>
+            struct PltfType<
+                exec::ExecGpuCudaRt<TDim, TSize, TKernelFnObj, TArgs...>>
+            {
+                using type = pltf::PltfCudaRt;
             };
         }
     }
